Add confirmation dialog

+++ AdvAnalyzer.WebApp/src/app/modules/search-query/containers/search-query-container/search-query-container.component.ts
@@ -3,6 +3,9 @@ import { PagedListQueryParams } from 'src/app/core/models/paged-list-query-param
 import { SearchQueryService } from '../../services/search-query.service';
 import { finalize, take } from 'rxjs';
 import { SearchQueryListComponent } from '../../components/search-query-list/search-query-list.component';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from 'src/app/shared/confirmation-dialog/confirmation-dialog.component';
+import { ConfirmationDialogModel } from 'src/app/shared/models/confirm-dialog.model';
 
 @Component({
   selector: 'app-search-query-container',
@@ -13,7 +16,7 @@ export class SearchQueryContainerComponent implements OnInit {
   public isLoading = false;
   @ViewChild(SearchQueryListComponent) searchQueryListComponent!: SearchQueryListComponent;
 
-  constructor(private readonly searchQueryService: SearchQueryService) { }
+  constructor(private readonly searchQueryService: SearchQueryService, public dialog: MatDialog) { }
 
   ngOnInit(): void {
   }
@@ -30,9 +33,24 @@ export class SearchQueryContainerComponent implements OnInit {
   }
 
   public delete(id: number) {
-    this.isLoading = true;
-    this.searchQueryService.delete(id).pipe(finalize(() => this.isLoading = false), take(1)).subscribe(x => {
-      this.loadData();
-    })
+    const message = `Are you sure you want to delete this Search Query?`;
+
+    const dialogData: ConfirmationDialogModel = {title: "Confirm Action", message: message};
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      maxWidth: "400px",
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      if(dialogResult) {
+        this.isLoading = true;
+        this.searchQueryService.delete(id).pipe(finalize(() => this.isLoading = false), take(1)).subscribe(x => {
+          this.loadData();
+        })
+      }
+    });
+
+
   }
 }
